refactor: move WebDriver setup to a dedicated module and update imports
