+++ modules/updater/updater.py
@@ -1,11 +1,10 @@
 import os
-import time
 
 from modules.settings import SAVE_HTML
 from modules.updater.data_processing.data_processor import set_filename_from_link
 from modules.updater.error_handler import fancy_error_handler
-from modules.updater.scraper.selenium_utils import scrape, setup_webdriver
 from modules.updater.sites.SiteFactory import SiteFactory
+from modules.updater.webdriver import setup_webdriver
 from modules.websites import search_links
 
 
@@ -17,7 +16,7 @@ def update_all_sites():
                 yield link_name
 
 
-@fancy_error_handler
+# @fancy_error_handler
 def update_sites_with_progress_bar(st):
     """Display progress bar while updating sites."""
     progress_bar, status_box = handle_update_progress(st)
@@ -45,17 +44,16 @@ def process_single_site(web_driver, search_link: str) -> None:
 def update_site(webdriver, search_link) -> str:
     """Download HTML content from the search link and save it to a file."""
     job_site = SiteFactory.identify_website(search_link)
-    search_block = scrape(webdriver, job_site)
+    search_block = job_site.scrape(webdriver)
 
     if SAVE_HTML:
-        filename = os.path.join(set_filename_from_link(search_link))
-        save_html_to_file(search_block, filename)
-
-    return filename if SAVE_HTML else ""
+        save_html_to_file(search_block, search_link)
+    return search_block
 
 
-def save_html_to_file(html_content: str, filename: str) -> None:
+def save_html_to_file(html_content: str, search_link: str) -> None:
     """Save HTML content to a file."""
+    filename = os.path.join(set_filename_from_link(search_link))
     with open(filename, "w", encoding="utf-8") as file:
         file.write(str(html_content))
 
@@ -65,5 +63,4 @@ def update_status(progress_bar, status_box, progress: int, link_name: str) -> No
     total = len(search_links)
     progress_bar.progress(progress / total)
     status_box.success(f"Downloaded: {link_name}")
-    time.sleep(3)
     status_box.empty()
